Stationary Ball option in GUI
Aatish, after working on p5 to make sure it ran OK, I just copy/pasted my new p5 code into this "Edit file" field-- is that how it works? Hope you can see the changes I made. If not, I introduced "isBallMoving" boolean variable and made a Stationary Ball option in the drop-down menu for interactions.

The one thing I couldn't figure out is in line 212-- currently, when you select Stationary Ball, the new stationary ball takes on the position of the Moving Ball before it stopped. I think it'd be better to have the position be reset to the center, but I wasn't sure how to go about doing that.

Alternatively, maybe it'd be nicer to have a checkbox for "moving" when the "Ball" option is selected from the dropdown menu-- in this scenario, I wouldn't mind the stationary ball taking on the old position. However, I wasn't sure how to code the GUI so that the checkbox would only be activated when a certain dropdown option is selected. Let me know what you think!